WIP: Use our native node addon in ImseService to import schemas. #24919
For now, this includes a crazy hack workaround for the fact that the
DgnDbNativeCode class (which we need to obtain a real handle to the open dgndb)
is not exported in @bentley/imodeljs-core/lib/IModel. Basically, I'm just using
webpack to inject the export into that module with exports-loader.

I realize this is a ludicrous thing to do - we need a real solution very soon.