Merged PR 522: Update to imodeljs-clients 5.0.0
Update to imodeljs-clients 5.0.0 after refactoring iModel Hub API in imodeljs-clients. This includes simplification of how codes are requested (we are hiding existence of MultiCodes as that is primarily optimization between Client API and iModel Hub).

As iModel Hub QA is not working properly today, I couldn't run tests again to make sure everything is still working. Tests were passing as of last week.