Setup support for long running open requests (TFS#891263)
Setup a way to execute long running async IModelConnection.open()
calls without timing out. The time out can happen especially when the
provisioner is setting up a brand new instance and downloading
a large iModel for the first time. The fixes build on the
infrastructure provided in the RpcInterface-s for pending requests.

The frontend issues an IModelConnection.open request to the backend
through the Rpc interface. The information on the IModel is passed as
an IModelToken - if the backend detects that it doesn't have a IModelDb
matching that token readily available, it will throw a "RpcPendingResponse"
exception, and the Rpc infrastructure repeats the call to the backend
after a delay. The delay is configured at the frontend - starts out at 500 ms,
and thereafter is doubled before every repeated call. This happens until
the total time exceeds a configurable timeout period. After that period,
an exception is thrown that the connection was timed out. The timeout can
be configured with IModelConnection.connectionTimeout, and is set to
a default of 5 minutes.

At the backend, once the request to open an IModelDb is received, a call to
setup the necessary IModelDb is initiated - the backend starts the process to
download the necessary IModelDb and ChangeSets from the iModelHub, and
merge them as necessary. The RpcPendingResponse exception is immediately
thrown back, and that causes a retry as described before.

If the request to open an iModel was already initialized by a previous call,
specific actions are taken based on the state of the pending request -
* if the IModelDb was successfully downloaded and opened, the backend returns the
token representing it to the frontend.
* if there was an error with opening the IModelDb, that's thrown back to
the frontend.
* if the setup of the IModelDb is still pending, an RpcPendingRequest exception
is thrown.

More details of the backend implementation:
The backend functionality is implemented with the help of a utility to
"memoize" long running async calls - this prevents duplicate calls
while the promise returned by previous calls is still pending, and
also allows to synchrnously check the resolution of these calls:
* A QueryablePromise wraps a promise returned by a long running function
call (that's to be memoized). The construct allows synchronous
query of the state of the wrapped promise.
* A generic memoizer caches these QueryablePromises in a map. The map
is keyed by the arguments made to the async function to me memoized.
* The above generic memoizer is customized to memoize calls to
IModelDb.open().
* A separate memoizer around BriefcaseManager.initCache()
prevents duplicate calls while the cache initialization
is in progress. (this part is WIP since the plan is to make initCache()
synchronous and independent of the accessToken).