Merged PR 926: Simplify installation of imodeljs-native-platform
Through the use if an npm install script, the addon now installs platform-specific packages automatically. It also provides a script for imodeljs-backend to use to load the appropriate addon. As a result, the dependencies of imodeljs-backend and the logic in NativePlatformRegistry are greatly simplified.
This moves us a step closer to installing node addons in the standard way.