Fix multiple constraint classes
This is a very pragmatic fix, but it does the trick for now. It's better to not perform the check than to fail whenever we have multiple constraint classes.

Since SchemaItemKey does no longer hold the type of schema item, we cannot bring the old check back. It seems doing something similar now would be more complex.

Related work items: #23156