Merged PR 6563: Move ec spec docs to imodeljs
This moves some of the ECObjects conceptual documentation from the [ec-spec](https://bentleycs.visualstudio.com/iModelTechnologies/iModelTechnologies/_git/ec-spec) repo into the [BIS section](https://imodeljs.github.io/iModelJs-docs-output/bis/) of the iModelJs SDK documentation site.

This will enable us to embed more relevant documentation in the iModel Schema Editor, and should help to explain (in more detail) several terms that are commonly referenced in the BIS docs.

Of course, this is really just a starting point.  There's still plenty of other ec-spec content that could be useful alongside the BIS docs, and we still have some work to do to improve the consistency between EC and BIS articles.  Bringing these pages in probably creates some redundancies in the existing BIS pages, and we'll also want to update the BIS docs to link to the EC docs when necessary.