Merged PR 9141: Refactor SettingsClient to accept OIDC access tokens
The current SettingsClient expects AuthorizationTokens (which are SAML only) and internally uses the SAML-only delegation service to obtain an AccessToken. This doesn't allow use of OIDC tokens. The API has therefore been changed so that it now takes AccessTokens already - which can come from OIDC or SAML. It is now up to the caller to provide the correct AccessToken from whatever path.

Related work items: #39638