Merged PR 9294: Create common classes to remove electron and express boilerplate
There's too much electron and express-specific boilerplate in our sample apps right now, and that's going to cause maintainability issues as we improve some of this basic configuration.

This is an attempt to move most of that into a few common classes in imodeljs-backend. In order to avoid requiring express and/or electron as dependencies of imodeljs-backend, I've made them devDependencies and used only their types.  So users will have to depend on (and pass) express or electron to use them, but this way imodeljs-backend never actually `require`s either itself.

Related work items: #26857