Merged PR 10111: Fixed settings client tests and iModelHub URL validation test
The settings client tests used an imodel which doesn't exist anymore. Change the test code to always use primary imodel for testing.

The URL validation test was broken as it did not filter non-hub URLs out anymore. Fixed that, by reusing some of the old validation code which had been removed. In particular, the URLs are logged into a separate file to simplify extraction a bit.