Merged PR 17557: i18n review and change to stripInternal setting
I reviewed and documented i18n for 1.0 readiness (as well as a few of the modules in imodeljs-frontend and imodeljs-clients)
While doing so, I marked some methods as @internal. That caused build problems because our setting for stripInternal was "true" in tsconfig-base.json. That causes @internal methods to be stripped out of the generated ".d.ts" files, which are used while building other packages. I believe setting stripInternal to "false" aligns with our usage of API Extractor and the guidelines we have circulated.