Merged PR 17367: Threading issue accessing Reality Data
There was a masive change of code by transfering the tile data, root document and other JSON composing the reality data to the Reality Data class of which an instance is specific to a single reality data. Prior to that a single RealityDataClient was used for most/all reality data which had to share cache data (the Azure blob address). Although some safeguard was temporarily added to minimize the chances of threading issues but it was not always safe and was inefficient. Now the cache data is specific to each data. Also added support for project (context) assignation into reality data so it needs not be given to access tile data and added a timestamp to the Azure blob address that gets refresh after 50 minutes (Reality Data Service only guarantees an hour bloob address validity)

Related work items: #83388