Merged PR 17835: Configurable backend tile request timeout; improved tile request logging
If a request for a tile tree or tile contents exceeds the timeout specified in IModelHost, a "pending" response will be returned to the front-end.
The default timeout for both trees and content is 20 seconds. A backend can override these values in IModelHost.startup().
Tile request log messages now include the tree + content ID.

Related work items: #84417, #84418