Merged PR 19784: Make sourcemaps available in NPM packages
This change adds the sourcemaps for the iModelJs external modules available in the npm packages. It increases the size of our npm packages, but allows a developer working on a package external to the imodeljs monorepo to see the iModelJs TypeScript source and set breakpoints in it. (The original TypeScript code is stored in the source map files).