Merged PR 20385: Update the extract-api script to work in an individual package without a devDep
The issue was it looked in the node_modules of the cwd for the .bin file and since each package did not have @microsoft/api-extractor as a dependency it could not be found in that pkgs bin folder.  Using require allows it to resolve the path into the build-tools node_modules which is where the bin directory exists.