Merged PR 27055: Presentation: added backend timing out handling
Added presentation backend request timing out after half a second by default. If presentation backend request times out, response code of 500 is returned to the frontend. If so, frontend handle this response with repeating the same request for 10 times, and if backend won't return success response, the PresentationError is thrown.

Related work items: #54944