Merged PR 27923: Fix high security vulnerabilities reported by rush audit
Updating css-loader to 2.1.1 and puppeteer to 1.13.0 seems to fix the current set of HIGH security vulnerabilities.

For css-loader, the jump was from 0.28.11 --> 2.1.1. This was what was dragging in the offending js-yaml.  @<Bill Goehrig> reviewed the css-loader changelog and recommended removing the use of the "minimize" option from the CSS configuration in our webpack scripts.

For puppeteer the move was to the current "chrome-stable" release.