Merged PR 28845: Fix issues with debugging certa tests
I finally found a way to make debugging these frontend-only certa tests a bit more reliable without writing a custom VSCode extension.

Basically, we just have to always start two debugging sessions even when we only care about attaching to the frontend.  This does make our launch.json a bit more complicated than before, but it removes the need for those preLaunch tasks that I think were breaking subsequent debugging sessions.

Since there are now a ton of launch configurations, and only the composites should be used, I've prefixed all the composites with ✔.  The "partial" configs (that won't work if started directly) are prefixed with ⚠.  This is a bit obnoxious, and I'd rather just hide the partial ones, but we'd need to publish an extension to do that.

This should also finally stop those unnecessary automatic breakpoints in the frontend - the tests will now simply wait for the debugger to resolve breakpoints before starting.  And the GPU inconsistencies should be fixed as well (no more tests that only fail in the debugger).