Merged PR 28994: Use fast-sass-loader fork and separate css files in prod
We are using fast-sass-loader (rather than sass-loader) but that lacked a way to pass the "compressed" outputStyle option to node-sass. I opened an issue two days ago with yibn2008/fast-sass-loader, but it hasn't even been looked at since, so I initially implemented it in a fork of fast-sass-loader. Doing that makes the size of our webpacked .js files much smaller because it removes the comments from the output css. However, rush does not cooperate with loading from an npm fork, so I had to abandon that option.

I also changed to using mini-css-extract-loader in production builds, so that the css files are separate from the javascript files after webpacking. That required changes to IModelJsLoader to load those files, and to buildIModelJs to put the output .css files into the correct lib/webserver directories. I added the optimize-css-assets-webpack-plugin to get the css files minimized for production (so the change to fast-sass-loader isn't needed for production).

The development modules (which contain the css since they don't run mini-css-extract-loader) are still too large because the css is not compressed. Hopefully the issue with fast-sass-loader will be addressed soon.