Merged PR 29269: Fix webpack/IModelJsLoader to avoid attempting load of .css files that don't exist
IModelJsLoader should not attempt to load .css files that don't exist on the server. Specifically, we don't separate out the .css files for development builds, so there is never a reason to try to load .css files in that environment. So I expanded the use of data-imjsversions string to have it list the required .css files as well as the javascript files.