Merged PR 29797: Add support for synchronously parsing strings into quantities. Add ability to show InputFieldMessage using NotificationManager.
Add methods parseIntoQuantityValue and loadFormatAndParsingMaps to QuantityFormatter to allow tools to both parse and format quantity values.
Add NotificationManager methods setInputFieldTypeDetails and closeInputFieldMessage to allow popup messages to be show adjacent to input field. This is used to show error message when unable to parse string into a quantity.
Update ui-test-app "ToolWithSettings" tool to demonstrate the parsing and formatting of length quantity using a toolsettings property specification.

Related work items: #127601