Merged PR 30276: fix declaration in internal addon api
This is a fix to the declaration of the interface NativeCrashReportingConfig within the *internal* IModelJsNative namespace.
The fix is to make the declaration match what the native code actually expects.
This is just tidying up. In fact, this interface declaration is not used. It is meant to document what the addon takes.
However, apps and the backend code actually use a public (alpha) class called IModelHost.CrashReportingConfig. That is a superset of IModelJsNative.NativeCrashReportingConfig. The backend code passes an instance of CrashReportingConfig to the addon to configure crash reporting. The object property names in CrashReportingConfig were already correct.