Merged PR 29943: Fix Bing Attribution Link
The Bing Attribution link was unreliable when there were elements behind it. All mouse events were coming to it, but it needed to stopPropagation to the canvas to prevent them passing to the canvas. Also changed our mouseenter/mouseleave to mouseover/mouseout to get events we could properly interpret as leaving and entering the canvas.