Merged PR 32493: Ensure disposal of disused tile trees
When the IModelConnection is closed, all tiles are disposed.
Periodically (at interval specified by application), we purge any tile trees that have not been drawn for at least that interval, AND are not currently in use by any ScreenViewport.
This ensures that, for example, if you switch from a spatial view to a sheet view, the tiles for all the spatial models displayed in the spatial view will eventually be discarded, freeing up their memory and WebGL resources.
Currently feature-gated; if application does not specify a purge interval then no purging will take place.

Related work items: #124612, #130061