Merged PR 33656: Support deferred processing in IModelTransformer
This PR implements a strategy for handling the case where the element being imported relies on other elements that haven't been imported yet.  One example case is trying to import a DrawingViewDefinition when the DrawingModel hasn't been imported yet. The strategy is as follows:

- Before importing, collect the set of *predecessor* elements for the specified element.
- If any of the *predecessor* elements have not yet been imported, skip the element
- Add a pass at the end to handle skipped elements

The new (currently @alpha) `Element.collectPredecessorIds` method (overridden at different levels in the Element class hierarchy) collects the ElementIds of *predecessor* elements (like parent, code scope, and modeledElement).  Those Element classes that override toJSON will also need to override collectPredecessorIds if the class requires another element (the predecessor) for itself to be inserted. For example, the category is a predecessor for GeometricElement.

This PR also includes some IModelTransformer logging improvements and other IModelTransformer refactoring. Since IModelTransformer is still @alpha, API changes to it are still allowed.