Merged PR 34598: Prototype version comparison
Added to the debug tools menu (under the big 'B') of display-test-app an "emulate version comparison" tool.
It expects to find a copy of the currently-opened iModel in same directory with ".rev" appended to its filename, to serve as the "revision" iModel.
Any elements present only in the revision iModel are treated as deletions (drawn in red, using tiles from the revision iModel); those present only in the primary iModel are treated as insertions (drawn in green); those in the selection set when the tool is initialized are treated as "modified" (drawn in blue); all others are treated as unchanged (non-locatable, grey, transparent).
Toggling of models, categories, view flags, etc should affect both the revision iModel and the primary iModel.
A subclass of SelectionTool is provided which enables locating elements in tiles belonging to the revision iModel.
Hilite, tooltips, etc "happen" to work in display-test-app because we do not use presentation rules / unified selection. A more highly-customized selection tool will be needed for the implementation of version comparison in Design Review.

Related work items: #145772