Merged PR 36631: New option to restrict tile cache URLs by client IP address.
Per discussion with the security team, this is a new option to restrict the signed tile cache access URLs that the backend generates to the client IP address. This is not enabled by default due to 1.0 and is an option since it could break clients in more complicated networking environments.