Merged PR 37269: Bug 148574: Fixed issue with opening iModels with names that are invalid on Unix
Bug 148574: Fixed issue with opening iModels with names that are invalid on Unix or Windows.

The fix involves changing all briefcase names to be "bc.bim" instead of "<iModelName>.bim". To avoid the old files sticking around in the cache folder, the cache version was updated -  this forces the entire cache folder to be recreated when the backend is initialized.

Related work items: #148574