Merged PR 39171: Build Plugins to Tar files
All files needed for a Plugin (the webpacked dev and prod versions, resources, localization files, etc.) are now built into a tar file.
Eventually, our production web server should be able to unpack the tar file and deliver the contents on a per-user basis. For now, the Express-based webserver in core/webserver does that, but it can only support a single user.
Plugins can be delivered not only by the web server delivering the iModelJs application, but by an external server (that support CORS requests).
Plugin tar files can optionally be signed.
These changes are needed to support the Safetibase plugin efforts.
NOTE: The public API for Plugin has been changed to accommodate the changes.