Merged PR 44261: Use "geometry GUID" in place of revision ID for tile cache differentiation
This allows cached tiles to be used across revisions as long as those revisions did not modify the geometry of the model.
Will only work for iModels which have been upgraded to the version of the BisCore schema that introduced the GeometricModel.GeometryGuid property - so basically none of them currently.

Related work items: #166625