Merged PR 44780: Implement browser test to reject IE and Old Edge
Our typescript compiled to javascript with the current tsconfig settings will not load at all in IE (parse error) and won't run at all in Old Edge.
It would be nice to reject those browsers in IModelJsLoader, but even that won't load (IE) or run (Old Edge), so we have to do it in a <script> in index.html. So every index.html needs to be updated with the code, a sample of which is in test-apps/ui-test-app/src/frontend/index.html. I put the same code in the documentation in docs/learning/frontend/BuildingIModelJsModules.md"