Merged PR 53865: Refactor tile traversal with preloaded parents for use by reality models.
The tile traversal for map tiles which preloads parents to improve the tile selection when the view frustum changes was previously used by map and terrain tiles and now is used also by reality models.  This reduces annoying dropouts to low resolution (or no geometry) when the view moves.      A seperate change was to default terrain lighting off as it dims terrain imagery display without a benefit in most situations.

Related work items: #185459