Merged PR 55344: UI: Refactor ControlledTree events
Refactored ControlledTree events to pass TreeNodeItem instead of nodeId. TreeModelNode might not be in the model if nodes where disposed while pending nodes were loaded for event, but data from TreeNodeItem might still be important for event handler.

This solves problem with unified selection hook not adding all selected node keys to selection because range selection event was initiated and parent node was collapsed. This led to situation where nodes for collapsed parent were loading but not inserted to model and event handler could not find them.

Related work items: #217011