Merged PR 57659: Align reality model coordinate system to IModel coordinates.
Reality models are defined in Earth Centric (ECEF) coordinates.  If a reality model is in the vicinity of an IModel but exported from a different origin then the curvature of the earth can cause misalignment (roll and pitch) between the reality and BIM model coordinates.  This causes the reality model data to be slightly misaligned and for large reality models (such as the EWR point clouds) the error can be significant.