Merged PR 58493: Fix and re-enable "direct" RPC tests
Most of these tests were failing because two copies of many imodeljs-common modules were being loaded: one in the webpack bundle and one from node_modules.  Webpacking the backend as well should avoid this.

Two of the failing tests ("should describe available RPC endpoints from the frontend" and "should reject a mismatched RPC interface request" actually needed to be refactored.  Basically, these tests implicitly expected that mutating an RpcInterface definition on the frontend would not affect that RpcInterface on the backend, but when we run in node, these are the same object reference.  To fix these, I've just refactored the tests to restore the original `interfaceVersion` on the interface definition **after** serializing and sending the test request, but **before** the backend handles the response.