Merged PR 60100: Updated display-test-app and display-performance-test-app to use OidcDesktopClient
Updated display-test-app and display-performance-test-app to use OidcDesktopClient.

This allows the refresh tokens to be stored in the key vault (Windows's Credential Vault or MacOS' Key Chain Store).
The token will now be valid for 30 days, and is now shared between all our internal test apps - so you login to one, you are logged into others also.
Signin is done in an external browser (outside the application).

Tested both the apps by re-signing in and stepping through the sign-in code.