Merged PR 60469: Language detector fix
We use i18nextBrowserLanguageDetector to detect the language that the user is displaying. However, the default for that is to cache the detected language in local storage, and then use that cached language on subsequent invocations. That traps the user into whatever language was first used when running the application. I changed the options to always ask the browser for the desired language.