Merged PR 61131: Remove the http2 tests from rpc integration tests as Node 12 does not support it
The implementation used for http2, Spdy, in the rpc integration tests does not work for Node 12 and is a known issue with the tool.  Since Express does not support http2 and the Spdy project has been abandoned as of Node 11, we'll just remove http2 testing for now.

The overall trend appears to be, do not use Node for http2 support but instead put something else in front of the it to add support.

https://github.com/expressjs/express/issues/2761
https://github.com/spdy-http2/handle-thing/pull/5#issue-229119787