Merged PR 63909: Refactor webpack-tools and create a separate, slimmer package for bundling just backends/agents
The `@bentley/webpack-tools` package has become a bit of a dumping ground, and as a result it pulls in a _ton_ of dependencies.  I think it needs to be broken into several smaller, more focused packages.

Bundling backends and agents with webpack should be much simpler than frontends, so we can start there and revisit how frontends are bundled later.

This change basically:
- Refactors webpack-tools to split out a separate, slimmer `@bentley/webpack-tools-core` package with common plugins and utils that can be used by _all_ iModel.js webpack builds.  These plugins should also make it easier for someone to configure webpack directly instead of using our build scripts.
    - I also re-implemented our webpack plugins in TypeScript to make them a bit more maintainable...

- Add a `@bentley/backend-webpack-tools` for bundling backends only.  This will be much lighter-weight than `@bentley/webpack-tools`, since it only handles backends, and _only_ wraps webpack - it does not handle linting/testing/etc.

Related work items: #268570