Merged PR 65872: Check for undefined or empty string in base name of tile cache url.
Check for undefined or empty string in base name of tile cache url.

On Safari neither of those values are allowed to be passed in as the
second parameter of URL().  Chrome and Firefox are more forgiving and
allow undefined (which is what we previously always passed).

There is a GitHub issue open for the Typescript type definition
(https://github.com/microsoft/TypeScript/issues/34725).