Merged PR 68146: Fixed hilite display for volume classifiers.
Hilite display of volume classifiers broke when the tile elision flag was changed to be on by default.  The reason is that the geometry is in a different form when this flag is on, so the checks in computeBatchHiliteRenderPass() were not seeing that it was a volume classifier.

Related work items: #270989