Merged PR 70276: Passing optional iModel context/project as it is the one to be used for Reality Data originating from ContextShare(RDS)
Added optional parameter to the class RealityModelTileClient for the iModel context (Project). Currently pre-attached Reality Data from Context Share use the original project. Bug was discovered as the referenced project did not exist anymore.
For the other non-preattached reality data the project was the same as the iModel so there is no change in behavior.