Merged PR 74170: Update build system
The following changes have been made within this PR:

- Remove the @bentley/webpack-tools package
- Remove the command line params from the Rush config, and remove the ignoreargs.js script from the build-tools pkg as it was only required because of the command line params.
- Update all iModel.js extensions to use the new webpacking strategy with the @bentley/extension-webpack-tools
- Change all iModel.js library package's build.  This includes,
   - Produce js and map files
   - No longer create "iModel.js Modules" in each pkg
   - Copy all assets explicitly into the lib folder
      -  This was previously done via the module system
   - All pseudolocalization is done via the script in the build-tools package.

- Last but not least, switch all test apps to react-scripts.  Details below...

### Updates to the test-apps

Since we're using CRA now for our frontend you can find a lot of details about CRA and how it works on their website [here](https://create-react-app.dev/).  Then more information in the [README](https://dev.azure.com/bentleycs/iModelTechnologies/_git/react-scripts?path=%2FREADME-bentley.md&version=GBbentley) for the Bentley/iModel.js fork of the react-scripts and why we need/want it.

Here I'd like to cover just the changes made to our current test-apps to make compatible with CRA/react-scripts and how that effects our dev workflow.

One of the main principles of CRA is that you have a `src` folder with a `index.ts` at the root, which is the entry point of an app, and a `public` folder with a `index.html` at the root.  Everything within the `src` folder is then subject to webpacking, including all of the assets that are parsed via loader (i.e. scss, css, json, etc.), and everything in the `public` folder is expected to live at the webroot when it's deployed so it's copied into the build output appropriately.

I added, and will continue to add more, about how to migrate from the previous setup to the new one within the NextVersion.md.

### How this changes our dev workflow

The old workflow was that you would need to rebuild the app every time you made a change within the app or any of the dependent packages.  Now with react-scripts you'll only need to ever rebuild the frontend if you are updating a dependency.

If you have an app, such as display-test-app running, and make a change in imodeljs-frontend all you need to do to consume that change in the app is run "rush build --to models-frontend" in a separate shell.  That should cause the app to automatically refresh within seconds to get the updated change.

### List of current issues to fix

- [x] ui-test-app in Electron isn't working...
- [x] Support svg sprites using svg-sprite-loader in the react-scripts fork.
- [x] come up with a new way to find the `imodeljs-config` repo.  Right now the apps look at their current directory for the config.
   - Have a new solution but would rather have all of it in a subsequent PR if possible.

Related work items: #285518, #285526