Remove imodeljs-webserver package - only used to serve frontends
The imodeljs-webserver used to be how we served out frontends for local testing but with the switch to react-scripts it comes with it's own version of the webserver.

In addition, none of our products use, or should really use, this as a way to serve out their frontend.  They're using more standard and easier ways of hosting a frontend.

This still leaves the @bentley/express-server which will continue to be the default way of hosting an iModel.js backend.