Merged PR 79645: Ensure 2d views have known extents at construction
Apps apparently like to fit the view immediately upon loading it. Previously the extents could not be computed until the tile tree was loaded.
For sheets, the model range is irrelevant - the extents are based on the sheet size.
For drawings, include the model range in the ViewStateProps.
Fixes "Maximum Window" error in PlantSight when loading and attempting to fit a drawing view, due to null range becoming maximal range.