Merged PR 76702: Thematic display: first stage support for smooth height gradients; introduction of thematic display API
This is the first stage of support for thematic display.  See the NextVersion.md entry for details on how to use it.

This includes the introduction of the thematic display API which will be expanded as needed.

This includes an addition to display-test-app which allows a user to enable and configure thematic display.

This first stage supports height mode with a smooth gradient.

EN: #289671
EN: #292396
EN: #292395
EN: #292394