Merged PR 80497: Tweak surface discard
We have logic in our shaders to prevent z-fighting in specific scenarios common to push-pull modeling: if some planar element is ~coincident with some non-planar surface (e.g., a shape sketched onto the surface of a slab), we discard the non-planar surface's pixels so that the planar geometry always draws on top.
Civil  models often represent roadways as layers of very-nearly-coincident surfaces.
Shaun discovered if he turns visible edges on and zooms out sufficiently, the edges of the layers underneath begin to show through the top layers, making the surface look black.
I tweaked the discard logic to specifically do:
-  IFF surface being drawn is a non-planar surface; and
-  IFF pick buffer contains planar geometry belonging to a different element:
-  THEN discard the non-planar surface pixel.

This resolves Shaun's issue and preserves the desired behavior.