Make IModelApp.startup() and IModelApp.shutdown() async.
Make `IModelApp.startup()` and `IModelApp.shutdown()` async.

We are currently ignoring the fact that some of the initialization steps in IModelApp are inherently asynchronous (specifically loading i10n namespaces, but there may be others).  While we may not have time to address all of this before 2.0, we should at least change the signatures of `IModelApp.startup()` and `IModelApp.shutdown()` now so we have the flexibility of fixing this in a future 2.x release without introducing another breaking change.  This will also give us a lot more flexibility to add other async initialization/termination logic.

Note that this is a **breaking public API change for 2.0**.

See !81175 for similar (backend) changes to `IModelHost`.