Merged PR 82453: Throttle TileTreeProps requests
Fixes bug with Shell iModel containing 1700+ models in one view - in web-based apps, the request for all of those models' TileTreeProps exceeds the browser limits for simultaneous http requests, so some tile trees do not display.

TileAdmin maintains a FIFO queue of requests for TileTreeProps. At most N requests may be in flight at any given time. The current in-flight requests are the first N entries in the queue.

When TileAdmin.requestTileTreeProps is called, it appends a new request to the queue, and if the queue length < N, dispatches it immediately.
When a request completes, throws an error, or is canceled, it is removed from the queue, and another request from the queue is dispatched.
When an IModelConnection is closed, any pending requests associated with that iModel are canceled.

Incidental changes: Moved Tiles and EditingFunctions out of IModelConnection.ts into their own separate files.