Merged PR 82061: Support thematic sensor display - first iteration
This is the first iteration of support for thematic display based on sensors.  This is an implementation of Inverse Distance Weighting to colorize a thematic display based on world position and associated values of an array of sensors.  See NextVersion.md for a description of how to use this new feature.

EN: #289684
EN: #285863