Merged PR 84050: Added shader debug / optimizaton tool
Added keyin command to display-test-app and code to extract both the glsl shader code and the translated Angle HLSL shader code, and output them to a folder (each shader in a separate file).  Also output a _makeShade.bat file to the folder which can compile the HLSL shaders into shader assembly (saved in .h files).  This can be used for study of shader code and optimization.

The translated hlsl file is further parsed and edited automatically to wire up the inputs and outputs so that it will compile properly.

New display-test-app keyin is "dta output shaders".
- Because of resource usage, ShaderProgram._debugShaders has to be compiled as true before this will do anything.
- In all cases only shaders that have been compiled so far are output.  Can use "ftd compile shaders" to get all compiled before using this.
- Keyin with no options outputs all hlsl and glsl, vertex and fragment shaders.  This can be filtered by using a string of letters as an option:
    V or F selects only Vertex or Fragment shaders (not specified is both)
    G or H selects only glsl or hlsl shaders (not specified is both)
    U or N selects only currently Used or Not-used shaders (not specified is all currently compiled shaders)
  e.g. "dta output shaders gu" will output all used glsl shaders, both fragment and vertex.
- The output folder defaults to d:\temp\shaders\ and will create the folder if it does not exist.  This path can be changed in code in outputShaders in Viewer.ts.
- A _OutputList.txt file is generated by the command that lists all the shaders that were output, and a flag of whether they are used or not.
- A _makeshader.bat is generated by the command to aid in compiling hlsl files.
    A /Zi option can be useful at times to add source information to the shader assembly.  See the fxc compiler options for more info.
    Running _makeShade (with or without compiler options) will compile all hlsl vertex shader files then all hlsl fragment shader files found in the folder, outputting the results into files of the same filename with a .h extension.  (.h was used because VS code syntax highlights this type well as a .h.)
Running this will also output a _ErrorFileList.txt that lists all files that had errors, followed by a build summary, as well as a _WarningsList.txt that contains any fxc compiler info on any warnings or errors encountered.