Merged PR 86239: cctl cache test was missing a call to close its db
It is doubly important to call ECDb.close, because that is the way that prepared statements in the statement cache are reset.
When ECDb.close is not called, the db is be closed when GC calls the native object's finalization callback. At that time, the native Db close will fail because statements are still active. You would only see this if you had SQLite logging enabled or in a debug build of the addon with asserts enabled.