Merged PR 91691: Fix Presentation onBeforeShutdown event listener.
Because the callback was set as just a reference to the `Presentation.terminate` function, it did not have the proper context for `this` when called as an  `IModelHost.onBeforeShutdown` event listener.  So calling `IModelHost.shutdown()` without first calling `Presentation.terminate()` was resulting in:

```
TypeError: Cannot read property '_clientsStorage' of undefined
      at terminate (node_modules\@bentley\presentation-backend\src\presentation-backend\Presentation.ts:143:14)
      at BeEvent.raiseEvent (node_modules\@bentley\bentleyjs-core\src\BeEvent.ts:98:26)
      at Function.shutdown (node_modules\@bentley\imodeljs-backend\src\IModelHost.ts:478:33)
      ...
```
Wrapping the callback in an arrow function should preserve the correct `this` and avoid those errors.