Merged PR 93161: Add custom activity id to potentially problem test
The iModel.js integration tests have an issue where one of the test suites is deleting all of the iModels.  Raman and I have tracked it down to the Super test user that is performing the deletes but have no idea which test suite it actually is.

After examining all of the tests that use the super user, these few are the ones that stand out the most.  Added a custom activity id to see if we can track it down.