Merged PR 93592: Support for multiple RPC interface clients and attaching interfaces after startup.
On the frontend, the new RpcRoutingToken can be used to identify an RPC configuration at initialization time and then subsequently used to retrieve the RpcInterface client that has the desired configuration.

On the backend, the new RpcConfiguration.attach method allows an interface to be injected into an active RPC configuration after startup. This can support the extension use case.