Merged PR 98304: Allow a graphic branch to selectively augment symbology overrides
Motivated by the need in editing workflows to display tiles containing modified or deleted elements and hide those elements in the context of those tiles while displaying them normally in other contexts.
Implemented semi-generically by allowing a FeatureSymbology.AppearanceProvider to be associated with a Branch.
Ultimately we expect a GeometricModelState to have access to the set of such elements such that it can implement FeatureSymbology.AppearanceProvider to simply check if the requested feature is associated with one of them.

Related work items: #386254