Merged PR 101284: Fix increase in number of requested tiles
2.3.0 changed the tile refinement logic to always sub-divide if any geometry had been omitted from the tile. This primarily targeted pathological data sets with stupidly huge project extents containing stupidly tiny geometry.
In sane data sets, it causes more tile requests than previously, for no gain in display fidelity. This was observed as an increase in the time required to load all of the tiles for the view in PlantSight and Design Review.
I kept the code for the 2.3.0 behavior but turned it off by default.