Merged PR 101789: AccessToken Changes to make extending AccessToken easier
Token.ts has been modified so that it is easier to extend AccessToken and still have any subclasses of AccessToken go through the flow of iModelApp / iModelHost properly. A class decorator called TokenPrefix has been created to attach to all Tokens so that a mapping of token prefix to constructor can be established. This mapping is used to generate the proper token type during the flow of authentication. Also moved BasicAccessToken to itwin-client package.