Merged PR 102217: Fix bug with AccessToken changes
Did not properly update static fromJson method to match the change in AccessToken's properties. Creating an accessToken from a json makes the assumption that a field _jwt exists which throws an error because I changed the _jwt field to be _tokenString to be more general for classes which extend AccessToken.