Merged PR 106698: Distinguish between 'surface is of textured type' and 'surface has texture'
Surface is of type `SurfaceType.TexturedLit` but its `texture` property is `undefined`.
Unclear how that happens - presumably either the texture image is not present in the tile, or cannot be deserialized? Can consistently reproduce it in Design Review; have not reproduced in display-test-app.
Is affecting at least one user (WSP) in production.

Related work items: #437906