Merged PR 108839: Fix tile usage tracking
We need to keep track of which tiles are currently in use by any viewport, so that we do not discard them.
Previously we used a shared set of "unique sets of viewports". It was pretty cute, but it had a fatal flaw in that the set of viewport sets could grow by one new entry every time we recreated the scene. Since every time we mark a tile as "in use", we must check that set to find the appropriate set to use, the bigger it grows, the slower that lookup should become.
SYNCHRO Modeler team observed such slowdown in production builds - the longer they navigated the view, the worse the performance became. Somehow the impact is much less noticeable in dev builds.

Replaced the "shared viewport sets" approach with something simpler.
Could have removed TileUsageMarker entirely except OrbitGT is weird (their real tile tree is hidden from us so we have to track usage of their graphics instead).
Incidentally removed unnecessary usage of custom container types where built-in maps/sets would suffice.

Related work items: #437921