Merged PR 110085: Force depth buffering for large ellipsoidal map tiles.  Correct imagery for  planar projection at max zoom.
When zooming out to near satellite level with depth buffering turned off, the tiles at the silhouette/horizon could produce depth anomalies so force them to depth buffer even if turned off.  These tiles are large enough that they will not be displayed at scale where underground BIM data is visible.  Unfortunately we still cannot avoid tile cracking with depth turned off as we cannot enable skirts and control their visibility in the non-depth buffered region.

Also allow level 1 imagery to be displayed.  This is never required with ellipsoidal projection as we limit the number of tiles to level 2 but for planar projection it is required and if not allowed then the number of tiles per graphic can exceed the max textures count.