Merged PR 112662: Fix locate for plan projections when map is drawn without depth.
When the map is drawn without depth buffer, **and** the model is a plan projection being drawn with display priority enabled, then although the model draws in front of the map, `readPixels` locates the map instead of the model.
This is because `SceneCompositor.drawForReadPixels` draws the background map as the first thing in the opaque pass, without depth - but it does so **after** we have drawn the opaque plan projection layers, so it overwrites them.
Change it so the background map is drawn as if it were an opaque plan projection layer itself, before any of the real plan projections.

Incidentally fix typo in Model and ModelState property `iSpatiallyLocated`.

Related work items: #442770