Merged PR 112950: Fix argument parsing
Bill added support for "quoted arguments" but did not provide for the ability to include a literal double-quote character within a quoted argument.
This broke key-ins that accept JSON, like `fdt apply view`, which I am trying to use for debugging.
I made it so that `"a literal "" character"` is parsed as `a literal " character`.