Merged PR 114162: Fix rotation problem by using calculate ECEF rather than relying on stored.
We were relying on correct stored ECEF transform to handle the up vector.  This can be miscalculated in different ways but we abandoned using it in favor of one calculated from GCS for map display a while ago.   Switch to using the calculated transform here - and also just use the default Z when we are within the project extents.