Merged PR 114538: Option to format rendering style JSON as quoted argument
Since we added support for quoted keyin arguments it is no longer possible to paste the output of `fdt save rendering style` directly as the input to `fdt apply rendering style` - you must replace all `"` characters with `""`. This is annoying, especially for users like Joe and Gino.
Added `quote` argument to the `save` keyin that does this replacement and encloses the result in quotes so it can be directly passed to the `apply` keyin.