Merged PR 115845: Cleaned up formatting of glsl shader code.
Mostly just a bunch of fixing/standardizing of new-lines and indentation in the shader code which gets produced.  I also sorted the variable declaration at the beginning so that you get the inputs first, followed by the globals, followed by the outputs.

This was purely just to make it easier to debug the shaders.

Related work items: #459998