Merged PR 116938: Ensure tile format version matches between backend and frontend
[Corresponding native PR](https://dev.azure.com/bentleycs/iModelTechnologies/_git/imodel02/pullrequest/116941)
Added IModelTileRpcInterface.queryVersionInfo().
Modified a different test so it doesn't have to be updated every time format version changes (on frontend).
I will create a small wiki article describing how and when to bump the tile format version. As part of this PR I'll update Caleb's new build script to output a link to that article on failure.