Merged PR 118294: #461166 - fix hilite/emphasize interaction
#461166 - fix hilite/emphasize interaction.

There were 2 basic problems with getting this to work correctly.  The first was that the hilite pass was drawing such that the last thing drawn won.  Thus if it was an emphasized element, it would overwrite the data for any hilited elements that were drawn before it.  Depth comparison is off for that pass, so it didn't matter which was actually in front of the other, just which drew last.  I solved this by turning on simple blending in the hilite pass so that now the hilite texture will contain all of the hilite regions AND all of the emphasized regions.

The second problem was with the composite shader.  Now that it had the information to be able to see the outline of the hilited region even when covered by an emphasized region, I had to re-work the code a little to make sure that the hilited outline was processed before doing any emphasizing.

Related work items: #461166