Merged PR 117920: Added color mix to thematic display for bg map terrain and point clouds.
This adds the ability to mix the color of the point cloud/bg map terrain main texture with the thematic color gradient color.

Note that for bg map terrain this takes away one of the secondary map layers for bg map terrain when only 8 textures are available (just when thematic display is used).  When more textures are available it uses an unused texture unit.