Merged PR 118873: Remove SpatialModelTileTrees
Deprecated as a result of [this PR](https://dev.azure.com/bentleycs/iModelTechnologies/_git/imodeljs/pullrequest/117687) but not removed because Design Review is using it for version comparison. That PR was cherry-picked into 2.7.x and didn't want to make DR react immediately. It was marked `@internal` so not otherwise concerned about API change.