Merged PR 119697: Fixed and modified hilite/emphasis interaction.
A previous fix let the hilite outline show through emphasized areas.  This fixes a problem that occurred when the hilite outline width is set to the non-default width of 2 pixels.  It also enables emphasis outlines to show through hilite areas and when the hidden ratios are set to more than 0.0 it will mix both the hilite color and the emphasis color in areas that are covered by both.  Also if the hilite and emphasis outlines are of different widths, then you will see both outlines around areas that are both hilited and emphasized.