Merged PR 120309: Add timer to panel auto expander
This PR adds a timer to panel auto expander. The timer is restarted when the mouse moves (over the expander) and is stopped when the mouse moves outside of the expander.
Grip auto expand functionality is unchanged (panel is expanded when mouse moves over it).

This PR also fixes an issue where collapsed panel is rendered as expanded (which leads to auto collapse functionality not actually collapsing the panel visually).

### Commits
- Fix issue where panel is rendered as not collapsed in collapsed state.
- Add timer to widget panel expander.
- Disable pointer events of center content.
- Fix transition issue when  resizing.
- Tests.
- Rush change.

Cherry picked from !119607