Merged PR 121302: Add new extension service modify scope to the Extension CLI
- Add the `imodel-extension-service:modify` scope to the Extension cli.
- Update the Extension client tests to reflect the current changes.

The `imodel-extension-service:modify` scope is going to be an internal-only scope which means the extension-cli is going to be the only publicly accessible way for anyone to create/manage/delete Extensions.

For a transition period, the current scope `imodel-extension-service-api` is going to support both read and write operations.  However, we will be turning that ability off shortly and the `imodel-extension-service-api` scope will only cover the read endpoints.

Once the switch happens, anyone who wants to create their own internal client that modifies the Extension Service will be required to get both `imodel-extension-service-api` and `imodel-extension-service:modify` scopes as they exclusively cover each set of endpoints, read and modification respectively.

Related work items: #422973