Merged PR 122160: Reduce threshold where GCS is used to calculate linear map transform.
When displaying background map we need a linear transform from the BIM coordinates to ECEF -- This is stored as the ECEF location --unfortunately this is not always valid and is not reset after being miscalculated in bridge.   In the case where there is a GCS we can calculate a valid transform - but currently this was only done if project more than 100KM from earth -- reduced this to 5KM -- It should probably be even tighter but will reserve that for future.