Merged PR 122923: Correct reality model location for iModels with GCS
The reality models were relying on correct ECEF transform in the case where there was a GCS. (The background map and terrain was already recalculating the transform from the GCS).

Remove problematic kludge to try and tilt reality models to match generation transform - this should not be necessary and caused more problems than it resolved.