Attach/detach view and viewport (#491)
Viewport.changeView() detaches the current view, if any, and attaches the new view. A ViewState can be attached to at most one Viewport at a time.

Attaching a view registers a bunch of event listeners with the ViewState and DisplayStyleState to keep the Viewport's state in synch with changes to them. This eliminates the need to remember to call special APIs to modify the view or style; and the need to manually invalidate the scene, render plan, etc after making such changes.
The ViewState is also notified when it is attached to a viewport so it can allocate any viewport-specific state (view attachments).

Detaching a view removes all of the event listeners registered when attaching and notifies the ViewState to deallocate any viewport-specific state (view attachments).

ModelSelectorState and CategorySelectorState now store their Ids in an ObservableSet. ViewState registers listeners on those sets, but *only* while attached to a Viewport - I didn't want to introduce unnecessary overhead there.

Incidental changes:
- The `equalState` methods of ModelSelectorState and CategorySelectorState were not implement correctly - they only compared equal if the sets of Ids were equal **by insertion order**. Fixed that.
- The tests for viewport changed events have been broken for a long time - they only looked like they were passing (exceptions were being caught and ignored).
- Viewport.ts had become a dumping ground. Split its contents into smaller files.
- Split some types out of ViewState.ts into their own files.

Remaining work for a subsequent PR: attaching/detaching context reality models, overriding their appearances, changing map imagery layers, and changing OSM building display all are mostly implemented in DisplayStyleState, but affect DisplayStyleSettings. We should move the logic to DisplayStyleSettings, add events for these properties, and have DisplayStyleState synchronize its own state in response to those events.